adds app component
